requirements.txt
absl-py==2.1.0
accelerate==1.2.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
dash==3.0.2
datasets==3.3.1
debugpy==1.8.11
decorator==5.1.1
deepspeed==0.16.1
defusedxml==0.7.1
diffusers==0.31.0
dill==0.3.8
docker-pycreds==0.4.0
dreamsim==0.2.1
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.11
fastjsonschema==2.21.1
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.55.8
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.70.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.33.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.3
notebook==7.3.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnxruntime-gpu==1.19.2
open3d==0.18.0
open_clip_torch==2.30.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow
platformdirs==4.3.6
plotly==6.0.1
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
pyequilib==0.5.8
pyglet==1.5.31
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.2.1
pyproj==3.6.1
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roma==1.5.1
rpds-py==0.22.3
ruff==0.9.9
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.14
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.12.0
torch==2.3.1
torchvision==0.18.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
trimesh==4.6.5
triton==2.3.1
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.0
wandb==0.19.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xformers
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
Pillow
huggingface_hub
onnxruntime-gpu